delete tokenizer folder;refactor the deployment test script to take external parameters
