+++ deploy/lmms_eval.sh
@@ -1,25 +1,115 @@
 #!/bin/bash
 
-# Set environment variables
-export CUDA_VISIBLE_DEVICES=0,1,2,3
-export PYTHON_MULTIPROCESSING_METHOD=spawn
-export VLLM_WORKER_MULTIPROC_METHOD=spawn
-export HF_HOME="~/.cache/huggingface"
-export HF_HUB_ENABLE_HF_TRANSFER="1"
-export NCCL_BLOCKING_WAIT=1
-export NCCL_TIMEOUT=18000000
-export NCCL_DEBUG=DEBUG
+usage() {
+    cat << EOF
+Usage: $0 [OPTIONS] <model_path1> <model_path2> ...
 
+Options:
+  -d, --devices DEVICES          CUDA devices to use (default: 0,1,2,3)
+  -t, --tensor-parallel SIZE     Tensor parallel size (default: 4)
+  -g, --gpu-memory UTILIZATION   GPU memory utilization (default: 0.9)
+  -r, --result-dir DIR           Base result directory (default: ./results)
+  -b, --batch-size SIZE          Batch size for auto tasks (default: auto)
+  -c, --hf-home                  Cache dir for HF datasets
+  --tasks TASK1,TASK2,...        Comma-separated list of tasks to evaluate (default: mmmu_val,docvqa_val,chartqa)
+  -h, --help                     Show this help message
+
+Examples:
+  bash $0 -d 0,1 -t 2 --gpu-memory 0.8 /path/to/model1 /path/to/model2
+  bash $0 --tasks mmmu_val,docvqa_val,chartqa /path/to/model1
+EOF
+}
+
+CUDA_VISIBLE_DEVICES="0,1,2,3"
 INFERENCE_TP_SIZE=4
+GPU_MEMORY_UTILIZATION=0.9
+RESULT_BASE_DIR="./results"
 BATCH_SIZE=16
-
+HF_HOME="~/.cache/huggingface"
 TASKS=("mmmu_val" "docvqa_val" "chartqa")
+
+POSITIONAL_ARGS=()
+
+while [[ $# -gt 0 ]]; do
+    case $1 in
+        -d|--devices)
+            CUDA_VISIBLE_DEVICES="$2"
+            shift 2
+            ;;
+        -t|--tensor-parallel)
+            INFERENCE_TP_SIZE="$2"
+            shift 2
+            ;;
+        -g|--gpu-memory)
+            GPU_MEMORY_UTILIZATION="$2"
+            shift 2
+            ;;
+        -r|--result-dir)
+            RESULT_BASE_DIR="$2"
+            shift 2
+            ;;
+        -b|--batch-size)
+            BATCH_SIZE="$2"
+            shift 2
+            ;;
+        -c|--hf-home)
+            HF_HOME="$2"
+            shift 2
+            ;;
+        --tasks)
+            IFS=',' read -ra TASKS <<< "$2"
+            shift 2
+            ;;
+        -h|--help)
+            usage
+            exit 0
+            ;;
+        -*|--*)
+            echo "Error: Unknown option: $1"
+            usage
+            exit 1
+            ;;
+        *)
+            POSITIONAL_ARGS+=("$1")
+            shift
+            ;;
+    esac
+done
+
+set -- "${POSITIONAL_ARGS[@]}"
 
 # Check if model paths are provided
 if [ $# -eq 0 ]; then
     echo "Usage: $0 <model_path1> <model_path2> ..."
     exit 1
 fi
+
+# Set environment variables
+export CUDA_VISIBLE_DEVICES=$CUDA_VISIBLE_DEVICES
+export PYTHON_MULTIPROCESSING_METHOD=spawn
+export VLLM_WORKER_MULTIPROC_METHOD=spawn
+export HF_HOME=$HF_HOME
+export HF_HUB_ENABLE_HF_TRANSFER="1"
+export NCCL_BLOCKING_WAIT=1
+export NCCL_TIMEOUT=18000000
+export NCCL_DEBUG=DEBUG
+
+echo "======================================================"
+echo "           Model Evaluation Configuration"
+echo "======================================================"
+echo "CUDA Visible Devices:      $CUDA_VISIBLE_DEVICES"
+echo "Tensor Parallel Size:      $INFERENCE_TP_SIZE"
+echo "GPU Memory Utilization:    $GPU_MEMORY_UTILIZATION"
+echo "Result Base Directory:     $RESULT_BASE_DIR"
+echo "Batch Size:                $BATCH_SIZE"
+echo "Tasks to Evaluate:         ${TASKS[*]}"
+echo "Number of Models:          $#"
+echo "Model Paths:"
+for model_path in "$@"; do
+    echo "  - $model_path"
+done
+echo "======================================================"
+echo
 
 # Iterate over all provided model paths
 for MODEL_PATH in "$@"; do
@@ -28,10 +118,9 @@ for MODEL_PATH in "$@"; do
     echo "======================================================"
     echo "Evaluating model: $MODEL_NAME"
     echo "Model path: $MODEL_PATH"
-    echo "======================================================"
 
     # Create dedicated result directory for the model
-    RESULT_PATH="./results/$MODEL_NAME"
+    RESULT_PATH="$RESULT_BASE_DIR/$MODEL_NAME"
     mkdir -p "$RESULT_PATH"
 
     for TASK in "${TASKS[@]}"; do
@@ -42,7 +131,7 @@ for MODEL_PATH in "$@"; do
         
         python3 -m lmms_eval \
             --model vllm \
-            --model_args model_version=$MODEL_PATH,gpu_memory_utilization=0.9,tensor_parallel_size=$INFERENCE_TP_SIZE \
+            --model_args model_version=$MODEL_PATH,gpu_memory_utilization=$GPU_MEMORY_UTILIZATION,tensor_parallel_size=$INFERENCE_TP_SIZE \
             --tasks $TASK \
             --batch_size $BATCH_SIZE \
             --log_samples \
